fix: change package name
Changed the package name of the app to fit accordingly to the conventions.
