+++ app/src/main/java/com/github/lookupgroup27/lookup/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.github.lookup.lookupapp
+package com.github.lookupgroup27.lookup
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -12,8 +12,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.semantics.testTag
 import androidx.compose.ui.tooling.preview.Preview
-import com.github.lookup.lookupapp.resources.C
-import com.github.lookup.lookupapp.ui.theme.SampleAppTheme
+import com.github.lookupgroup27.lookup.resources.C
+import com.github.lookupgroup27.lookup.ui.theme.SampleAppTheme
 
 class MainActivity : ComponentActivity() {
   override fun onCreate(savedInstanceState: Bundle?) {
