+++ app/src/main/java/com/github/lookupgroup27/lookup/ui/theme/Theme.kt
@@ -1,4 +1,4 @@
-package com.github.lookup.lookupapp.ui.theme
+package com.github.lookupgroup27.lookup.ui.theme
 
 import android.app.Activity
 import android.os.Build
