+++ app/src/main/java/com/github/lookupgroup27/lookup/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.github.lookup.lookupapp.ui.theme
+package com.github.lookupgroup27.lookup.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
