+++ app/src/test/java/com/github/lookupgroup27/lookup/screen/SecondScreen.kt
@@ -1,7 +1,7 @@
-package com.github.lookup.lookupapp.screen
+package com.github.lookupgroup27.lookup.screen
 
 import androidx.compose.ui.test.SemanticsNodeInteractionsProvider
-import com.github.lookup.lookupapp.resources.C
+import com.github.lookupgroup27.lookup.resources.C
 import io.github.kakaocup.compose.node.element.ComposeScreen
 import io.github.kakaocup.compose.node.element.KNode
